portfolio/requirements.txt
Django==5.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
whitenoise==6.8.2
requests==2.31.0

